Semestr_4/Jezyki_skryptowe/projekt/requirements.txt
tkinterdnd2~=0.4.3
PyPDF2~=3.0.1
python-docx~=1.2.0
pytest~=8.4.1